+++ Greg.Xrm.Command.Core/Commands/Script/ScriptAllCommandExecutor.cs
@@ -7,7 +7,7 @@ namespace Greg.Xrm.Command.Commands.Script
     public class ScriptAllCommandExecutor : ICommandExecutor<ScriptAllCommand>
     {
         private readonly IScriptExtractionService extractionService;
-        public ScriptAllCommandExecutor(IOutput output, IOrganizationServiceRepository organizationServiceRepository, IScriptExtractionService extractionService)
+        public ScriptAllCommandExecutor(IScriptExtractionService extractionService)
         {
             this.extractionService = extractionService;
         }
